refactor(islamic-website): rename utils files, correct returns
