+++ apps/islamic-website/utils/get-disclaimer.ts
