+++ apps/islamic-website/utils/get-fatwa.ts
