+++ apps/islamic-website/utils/get-privacy-policy.ts
@@ -1,13 +1,13 @@
-import { getWhitepaperData } from '@haqq/data-access-falconer';
+import { getIslamicPrivacyPolicy } from '@haqq/data-access-falconer';
 import { REVALIDATE_TIME } from '../constants';
 import { cache } from 'react';
 
 export const revalidate = REVALIDATE_TIME;
 
-export const getWhitepaperContentFromFalconer = cache(
+export const getPrivacyPolicyContentFromFalconer = cache(
   async (locale: string) => {
     try {
-      const data = await getWhitepaperData(
+      const data = await getIslamicPrivacyPolicy(
         {
           next: {
             revalidate,
@@ -16,9 +16,11 @@ export const getWhitepaperContentFromFalconer = cache(
         locale,
       );
 
-      return data.wp;
+      return data;
     } catch (error) {
       console.error(error);
     }
+
+    return undefined;
   },
 );
