+++ apps/islamic-website/utils/get-whitepaper.ts
@@ -1,13 +1,13 @@
-import { getPrivacyPolicyData } from '@haqq/data-access-falconer';
+import { getIslamicWhitepaper } from '@haqq/data-access-falconer';
 import { REVALIDATE_TIME } from '../constants';
 import { cache } from 'react';
 
 export const revalidate = REVALIDATE_TIME;
 
-export const getPrivacyPolicyContentFromFalconer = cache(
+export const getWhitepaperContentFromFalconer = cache(
   async (locale: string) => {
     try {
-      const data = await getPrivacyPolicyData(
+      const data = await getIslamicWhitepaper(
         {
           next: {
             revalidate,
@@ -16,9 +16,11 @@ export const getPrivacyPolicyContentFromFalconer = cache(
         locale,
       );
 
-      return data.pp;
+      return data;
     } catch (error) {
       console.error(error);
     }
+
+    return undefined;
   },
 );
